+++ src/SecureCore.Auth.OAuth/Extensions/OAuthBuilder.cs
@@ -6,12 +6,7 @@ namespace SecureCore.Auth.OAuth.Extensions;
 /// <summary>
 /// Permite configurar los proveedores y opciones del ecosistema OAuth.
 /// </summary>
-public class OAuthBuilder
+public class OAuthBuilder(IServiceCollection services)
 {
-    public IServiceCollection Services { get; }
-
-    public OAuthBuilder(IServiceCollection services)
-    {
-        Services = services;
-    }
+    public IServiceCollection Services { get; } = services;
 }